requirements.txt
django<=4.1.0,<4.2.0
djangorestframework
Pillow
drf-spectacular